requirements.txt
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
